Move searcher to lib
