+++ employee_directory/employee_directory.rb
@@ -2,16 +2,6 @@ require 'data_mapper'
 require_relative 'seeder'
 
 module EmployeeDirectory
-  class Searcher
-    def search(employee_reference)
-      if employee_reference =~ /^\d*$/
-        [Employee.get(employee_reference)]
-      else
-        Employee.all(:name.like => "%#{employee_reference}%")
-      end
-    end
-  end
-
   module_function
 
   def init(database_url)
