+++ lib/searcher.rb
@@ -1,0 +1,9 @@
+class Searcher
+  def search(employee_reference)
+    if employee_reference =~ /^\d*$/
+      [Employee.get(employee_reference)]
+    else
+      Employee.all(:name.like => "%#{employee_reference}%")
+    end
+  end
+end
